[Fix] UI navigation [Add] Binding for setting focused UI element
