+++ engine/bindings/private/game/game_bindings.cpp
@@ -2,28 +2,19 @@
 
 #include "cheats_component.hpp"
 #include "components/rigidbody_component.hpp"
-#include "components/static_mesh_component.hpp"
-#include "components/transform_component.hpp"
-#include "components/transform_helpers.hpp"
-#include "cpu_resources.hpp"
 #include "ecs_module.hpp"
 #include "entity/wren_entity.hpp"
 #include "game_module.hpp"
 #include "model_loading.hpp"
 #include "physics/shape_factory.hpp"
 #include "physics_module.hpp"
-#include "renderer_module.hpp"
 #include "systems/lifetime_component.hpp"
 #include "ui/game_ui_bindings.hpp"
-
-#include <spdlog/fmt/fmt.h>
+#include "ui_module.hpp"
 
 #include <Jolt/Physics/Collision/Shape/CapsuleShape.h>
 
-#include "ui_progress_bar.hpp"
 #include "ui_text.hpp"
-
-#include <ui_image.hpp>
 
 namespace bindings
 {
@@ -75,80 +66,11 @@ void AlterPlayerHeight(MAYBE_UNUSED GameModule& self, PhysicsModule& physicsModu
     }
 }
 
-std::optional<std::shared_ptr<HUD>> GetHUD(GameModule& self)
+void SetGamepadActiveButton(UIModule& self, std::shared_ptr<UIElement> button)
 {
-    if (auto lock = self._hud.lock())
-    {
-        return lock;
-    }
-    return std::nullopt;
+    self.uiInputContext.focusedUIElement = button;
 }
 
-void UpdateHealthBar(HUD& self, const float health)
-{
-    if (auto locked = self.healthBar.lock(); locked != nullptr)
-    {
-        locked->SetFractionFilled(health);
-    }
-}
-
-void UpdateAmmoText(HUD& self, const int ammo, const int maxAmmo)
-{
-    if (auto locked = self.ammoCounter.lock(); locked != nullptr)
-    {
-        locked->SetText(std::to_string(ammo) + "/" + std::to_string(maxAmmo));
-    }
-}
-
-void UpdateUltBar(HUD& self, const float ult)
-{
-    if (auto locked = self.ultBar.lock(); locked != nullptr)
-    {
-        locked->SetFractionFilled(ult);
-    }
-}
-
-void UpdateScoreText(HUD& self, const int score)
-{
-    if (auto locked = self.scoreText.lock(); locked != nullptr)
-    {
-        locked->SetText(std::string("Score: ") + std::to_string(score));
-    }
-}
-
-void UpdateMultiplierText(HUD& self, const float multiplier)
-{
-    if (auto locked = self.multiplierText.lock(); locked != nullptr)
-    {
-        locked->SetText(fmt::format("{:.1f}", multiplier).append("x"));
-    }
-}
-
-void UpdateGrenadeBar(HUD& self, const float charge)
-{
-    if (auto locked = self.grenadeBar.lock(); locked != nullptr)
-    {
-        locked->SetFractionFilled(charge);
-    }
-}
-
-void UpdateDashCharges(HUD& self, int charges)
-{
-    for (int32_t i = 0; i < static_cast<int32_t>(self.dashCharges.size()); i++)
-    {
-        if (auto locked = self.dashCharges[i].lock(); locked != nullptr)
-        {
-            if (i < charges) // Charge full
-            {
-                locked->display_color = glm::vec4(1);
-            }
-            else // Charge empty
-            {
-                locked->display_color = glm::vec4(1, 1, 1, 0.2);
-            }
-        }
-    }
-}
 }
 
 void BindGameAPI(wren::ForeignModule& module)
@@ -165,17 +87,13 @@ void BindGameAPI(wren::ForeignModule& module)
 
     game.func<&GameModule::GetMainMenu>("GetMainMenu");
     game.func<&GameModule::SetMainMenuEnabled>("SetMainMenuEnabled");
+
     game.func<&GameModule::SetHUDEnabled>("SetHUDEnabled");
-    BindMainMenu(module);
+    game.func<&GameModule::GetHUD>("GetHUD");
 
-    game.funcExt<bindings::GetHUD>("GetHUD");
-    auto& hud = module.klass<HUD>("HUD");
+    auto& ui = module.klass<UIModule>("UIModule");
+    auto& element = module.klass<UIElement>("UIElement");
+    ui.funcExt<bindings::SetGamepadActiveButton>("SetSelectedElement");
 
-    hud.funcExt<bindings::UpdateHealthBar>("UpdateHealthBar", "Update health bar with value from 0 to 1");
-    hud.funcExt<bindings::UpdateAmmoText>("UpdateAmmoText", "Update ammo bar with a current ammo count and max");
-    hud.funcExt<bindings::UpdateUltBar>("UpdateUltBar", "Update ult bar with value from 0 to 1");
-    hud.funcExt<bindings::UpdateScoreText>("UpdateScoreText", "Update score text with score number");
-    hud.funcExt<bindings::UpdateGrenadeBar>("UpdateGrenadeBar", "Update grenade bar with value from 0 to 1");
-    hud.funcExt<bindings::UpdateDashCharges>("UpdateDashCharges", "Update dash bar with number of remaining charges");
-    hud.funcExt<bindings::UpdateMultiplierText>("UpdateMultiplierText", "Update multiplier number");
+    BindGameUI(module);
 }
