+++ engine/bindings/private/game/ui/game_ui_bindings.hpp
@@ -1,4 +1,4 @@
 #pragma once
 #include "wren_common.hpp"
 
-void BindMainMenu(wren::ForeignModule& module);
+void BindGameUI(wren::ForeignModule& module);
